+++ src/ui/main/disable_telemetry.rs
@@ -9,8 +9,11 @@ use super::{App, AppMsg};
 pub fn disable_telemetry(sender: ComponentSender<App>) {
     sender.input(AppMsg::DisableButtons(true));
 
+    let config = Config::get().unwrap();
+
     std::thread::spawn(move || {
-        let telemetry = TELEMETRY_SERVERS
+        let telemetry = config.launcher.edition
+            .telemetry_servers()
             .iter()
             .map(|server| format!("echo '0.0.0.0 {server}' >> /etc/hosts"))
             .collect::<Vec<String>>()