+++ CHANGELOG.md
@@ -7,6 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added multi-region support
+
+### Fixed
+
+- Fixed logo size in the first run window
+- Fixed 7z dependency check on some systems
+
+### Changed
+
+- Updated Turkish
+- Updated Italian
+- Updated Japanese
+- Updated Turkish
+- Updated Indonesian
+
 ## [1.3.0] - 02.08.2023
 
 ### Added